pyproject.toml
[project]
name = "notifications-sprint-1"
version = "0.1.0"
description = "Notification service"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.14.1",
    "asyncpg>=0.30.0",
    "croniter>=6.0.0",
    "fastapi[standard]>=0.115.8",
    "orjson>=3.10.15",
    "psycopg>=3.1.18",
    "pydantic-settings>=2.7.1",
    "pyjwt[crypto]>=2.10.1",
    "redis>=5.2.1",
    "sentry-sdk>=2.20.0",
    "sqlalchemy>=2.0.38",
]

[dependency-groups]
dev = [
    "isort>=6.0.0",
    "mypy>=1.15.0",
    "pre-commit>=4.1.0",
    "ruff>=0.9.5",
    "sqlalchemy[mypy]>=2.0.38",
]

[tool.ruff]
exclude = [".venv", "./notification_service/src/db/migrations/"]
line-length = 119
indent-width = 4
target-version = "py312"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "C",   # flake8-comprehensions
    "B",   # flake8-bugbear
    "UP",  # pyupgrade
    "RUF", # ruff-specific rules
    "TID", # flake8-tidy-imports
    "TCH", # flake8-type-checking
    "PTH", # flake8-use-pathlib
    "PL"   # pylint
]
ignore = [
    "RUF002",  # Игнорируем проверку кириллицы в докстрингах
    "RUF003",  # Игнорируем проверку кириллицы в комментариях
    "I001"     # Игнорируем проверку сортировки импортов, так как используем isort
]
mccabe = { max-complexity = 10 }
pydocstyle = { convention = "google" }
pylint = { max-args = 6 }

[tool.ruff.lint.isort]
combine-as-imports = true
force-sort-within-sections = true
known-first-party = ["api", "core", "db", "documents", "exceptions", "middlewares", "services", "main", "handlers", "models", "domain", "repositories"]
lines-after-imports = 2
relative-imports-order = "closest-to-furthest"
split-on-trailing-comma = true

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["E402"]

[tool.mypy]
python_version = "3.12"
ignore_missing_imports = true
disallow_untyped_calls = true
disallow_untyped_defs = true
plugins = [
  "sqlalchemy.ext.mypy.plugin"
]

[tool.isort]
profile = "black"
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 119
skip = [".venv", "notification_service/src/db/migrations/"]
combine_as_imports = true
sections = ["STDLIB", "THIRDPARTY", "FIRSTPARTY", "LOCALFOLDER"]
known_first_party = ["src"]
known_third_party = ["sqlalchemy", "pydantic", "asyncpg", "croniter", "alembic"]
import_heading_stdlib = "stdlib"
import_heading_thirdparty = "thirdparty"
import_heading_firstparty = "project"
